ios portion

+++ Schema/uWaterlooMath136Score.py
@@ -1,5 +1,5 @@
 from __future__ import print_function
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String,Float
 from sqlalchemy.ext.declarative import declarative_base
 
 
@@ -8,7 +8,7 @@ class uWaterlooMath136Score(declarative_base()):
     __tablename__ = 'uwaterloomath136score'
 
     fileName = Column(String, primary_key=True)
-    score = Column(Integer)
+    score = Column(Float)
 
     def __init__(self, fileName, score):
         self.fileName = fileName
